requirements.txt
altgraph==0.17.3
commonmark==0.9.1
cryptocode==0.1
future==0.18.2
pefile==2022.5.30
Pillow==9.2.0
pycryptodome==3.15.0
pycryptodomex==3.15.0
Pygments==2.13.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.11
pywin32-ctypes==0.2.0
rich==12.6.0
